refactor: update main components name for better readability
