+++ src/components/Contacts.jsx
@@ -1,11 +1,11 @@
-import Button from "./Button";
+// import Button from "./Button";
 import Heading from "./Heading";
 import Section from "./Section";
 import { curve } from "../assets";
-import Tagline from "./Tagline";
-import { roadmap } from "../constants";
-import { check2, grid, loading1 } from "../assets";
-import { Gradient } from "./design/Roadmap";
+// import Tagline from "./Tagline";
+// import { roadmap } from "../constants";
+// import { check2, grid, loading1 } from "../assets";
+// import { Gradient } from "./design/Roadmap";
 
 const Roadmap = () => (
   <Section className="overflow-hidden" id="roadmap">
