+++ src/components/TeamsList.jsx
@@ -1,14 +1,10 @@
 import Section from "./Section";
 import { smallSphere, stars } from "../assets";
 import Heading from "./Heading";
-import PricingList from "./PricingList";
-import { LeftLine, RightLine } from "./design/Pricing";
 import TeamCard from "./TeamCard";
 import { mentors, teamleads, contributors } from '../constants/team.json';
 import '../../public/css/TeamCard.css';
 import '../../public/css/Team.css';
-
-
 
 
 const Pricing = () => {
